# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FPLense API",
    "summary": "FPL points forecasts, squad optimizer, season replay and team tools.",
    "description": "Prices are integer tenths of £1m (55 = £5.5m). Expected points come from a LightGBM model validated walk-forward; hindsight-best squads are the best legal £100m squads on actual points (not FPL's Dream Team). Not affiliated with the Premier League.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/squads/best": {
      "get": {
        "tags": [
          "squads"
        ],
        "summary": "Best",
        "description": "The model's pick: the frozen archive for a past GW, the live best team for the next.",
        "operationId": "best_api_squads_best_get",
        "parameters": [
          {
            "name": "gw",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 38,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "default: the next gameweek",
              "title": "Gw"
            },
            "description": "default: the next gameweek"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Squad"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/squads/optimize": {
      "post": {
        "tags": [
          "squads"
        ],
        "summary": "Optimize",
        "description": "Re-solve the squad ILP for the next GW with your budget, horizon, bench weight, locks\nand bans. An impossible lock set answers 422 naming the broken rule.",
        "operationId": "optimize_api_squads_optimize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Squad"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/squads/evaluate": {
      "post": {
        "tags": [
          "squads"
        ],
        "summary": "Evaluate",
        "description": "Expected (and, for a finished GW, actual) points, cost and a plain-English rule report\nfor any 0-15 players; a full squad without an XI gets the lineup helper's.",
        "operationId": "evaluate_api_squads_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/transfers/plan": {
      "post": {
        "tags": [
          "transfers"
        ],
        "summary": "Plan",
        "description": "Best moves for T = 0..5 transfers (net of -4 hits beyond your free ones) and a\nstep-by-step path from your squad to the model's best squad.",
        "operationId": "plan_api_transfers_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EvaluateIn": {
        "properties": {
          "players": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 15,
            "title": "Players"
          },
          "starters": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array",
                "maxItems": 11
              },
              {
                "type": "null"
              }
            ],
            "title": "Starters"
          },
          "captain": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captain"
          },
          "vice": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vice"
          },
          "gw": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gw"
          }
        },
        "type": "object",
        "required": [
          "players"
        ],
        "title": "EvaluateIn"
      },
      "EvaluateOut": {
        "properties": {
          "gw": {
            "type": "integer",
            "title": "Gw"
          },
          "finished": {
            "type": "boolean",
            "title": "Finished"
          },
          "legal": {
            "type": "boolean",
            "title": "Legal"
          },
          "problems": {
            "items": {
              "$ref": "#/components/schemas/Problem"
            },
            "type": "array",
            "title": "Problems"
          },
          "cost": {
            "type": "integer",
            "title": "Cost"
          },
          "bank": {
            "type": "integer",
            "title": "Bank"
          },
          "starters": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Starters"
          },
          "bench": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Bench"
          },
          "captain": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captain"
          },
          "vice": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vice"
          },
          "expected_points": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Points"
          },
          "horizon_expected": {
            "items": {
              "$ref": "#/components/schemas/GwPoints"
            },
            "type": "array",
            "title": "Horizon Expected"
          },
          "actual_points": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Points"
          },
          "players": {
            "items": {
              "$ref": "#/components/schemas/SquadPlayer"
            },
            "type": "array",
            "title": "Players"
          }
        },
        "type": "object",
        "required": [
          "gw",
          "finished",
          "legal",
          "problems",
          "cost",
          "bank",
          "starters",
          "bench",
          "captain",
          "vice",
          "expected_points",
          "horizon_expected",
          "actual_points",
          "players"
        ],
        "title": "EvaluateOut"
      },
...
      "GwPoints": {
        "properties": {
          "gw": {
            "type": "integer",
            "title": "Gw"
          },
          "points": {
            "type": "number",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "gw",
          "points"
        ],
        "title": "GwPoints"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Move": {
        "properties": {
          "out": {
            "$ref": "#/components/schemas/SquadPlayer"
          },
          "into": {
            "$ref": "#/components/schemas/SquadPlayer"
          },
          "gain": {
            "type": "number",
            "title": "Gain",
            "description": "P_h of the player in minus the player out"
          }
        },
        "type": "object",
        "required": [
          "out",
          "into",
          "gain"
        ],
        "title": "Move"
      },
      "OptimizeIn": {
        "properties": {
          "budget": {
            "type": "integer",
            "maximum": 1500.0,
            "minimum": 500.0,
            "title": "Budget",
            "description": "tenths of £1m",
            "default": 1000
          },
          "horizon": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Horizon",
            "default": 5
          },
          "bench_weight": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Bench Weight",
            "default": 0.1
          },
          "locked": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 15,
            "title": "Locked"
          },
          "banned": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 200,
            "title": "Banned"
          }
        },
        "type": "object",
        "title": "OptimizeIn"
      },
      "PathOut": {
        "properties": {
          "target": {
            "$ref": "#/components/schemas/Squad"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/PathStepOut"
            },
            "type": "array",
            "title": "Steps"
          },
          "make_now": {
            "type": "integer",
            "title": "Make Now"
          },
          "advice": {
            "type": "string",
            "title": "Advice"
          }
        },
        "type": "object",
        "required": [
          "target",
          "steps",
          "make_now",
          "advice"
        ],
        "title": "PathOut"
      },
      "PathStepOut": {
        "properties": {
          "n_moves": {
            "type": "integer",
            "title": "N Moves"
          },
          "new_moves": {
            "items": {
              "$ref": "#/components/schemas/Move"
            },
            "type": "array",
            "title": "New Moves"
          },
          "gain_before_hits": {
            "type": "number",
            "title": "Gain Before Hits"
          },
          "hits": {
            "type": "integer",
            "title": "Hits"
          },
          "net_gain": {
            "type": "number",
            "title": "Net Gain"
          },
          "bank_after": {
            "type": "integer",
            "title": "Bank After"
          }
        },
        "type": "object",
        "required": [
          "n_moves",
          "new_moves",
          "gain_before_hits",
          "hits",
          "net_gain",
          "bank_after"
        ],
        "title": "PathStepOut"
      },
...
      "Problem": {
        "properties": {
          "rule": {
            "type": "string",
            "title": "Rule"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "rule",
          "message"
        ],
        "title": "Problem"
      },
...
      "Squad": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "gw": {
            "type": "integer",
            "title": "Gw"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "players": {
            "items": {
              "$ref": "#/components/schemas/SquadPlayer"
            },
            "type": "array",
            "title": "Players"
          },
          "starters": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Starters"
          },
          "bench": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Bench"
          },
          "captain": {
            "type": "integer",
            "title": "Captain"
          },
          "vice": {
            "type": "integer",
            "title": "Vice"
          },
          "cost": {
            "type": "integer",
            "title": "Cost"
          },
          "bank": {
            "type": "integer",
            "title": "Bank"
          },
          "budget": {
            "type": "integer",
            "title": "Budget"
          },
          "expected_points": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Points",
            "description": "XI + captain for the GW"
          },
          "horizon_expected": {
            "items": {
              "$ref": "#/components/schemas/GwPoints"
            },
            "type": "array",
            "title": "Horizon Expected"
          },
          "actual_points": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Points",
            "description": "XI after auto-subs + captain, if finished"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code",
            "description": "shareable team code"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "problems": {
            "items": {
              "$ref": "#/components/schemas/Problem"
            },
            "type": "array",
            "title": "Problems",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "kind",
          "gw",
          "label",
          "players",
          "starters",
          "bench",
          "captain",
          "vice",
          "cost",
          "bank",
          "budget",
          "expected_points",
          "horizon_expected",
          "actual_points",
          "code"
        ],
        "title": "Squad"
      },
      "SquadPlayer": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "web_name": {
            "type": "string",
            "title": "Web Name"
          },
          "pos": {
            "type": "string",
            "enum": [
              "GK",
              "DEF",
              "MID",
              "FWD"
            ],
            "title": "Pos"
          },
          "club": {
            "type": "string",
            "title": "Club"
          },
          "club_short": {
            "type": "string",
            "title": "Club Short"
          },
          "price": {
            "type": "integer",
            "title": "Price"
          },
          "expected": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected"
          },
          "P_h": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P H"
          },
          "actual": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual"
          },
          "minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minutes"
          },
          "starter": {
            "type": "boolean",
            "title": "Starter"
          },
          "bench_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bench Order"
          },
          "captain": {
            "type": "boolean",
            "title": "Captain"
          },
          "vice": {
            "type": "boolean",
            "title": "Vice"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "rating_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating Source"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "news": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "News"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "web_name",
          "pos",
          "club",
          "club_short",
          "price",
          "expected",
          "P_h",
          "actual",
          "minutes",
          "starter",
          "bench_order",
          "captain",
          "vice",
          "photo_url",
          "rating",
          "rating_source",
          "status",
          "news"
        ],
        "title": "SquadPlayer"
      },
...
      "TransferIn": {
        "properties": {
          "team_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          },
          "squad": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array",
                "maxItems": 15,
                "minItems": 15
              },
              {
                "type": "null"
              }
            ],
            "title": "Squad"
          },
          "bank": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Bank",
            "description": "tenths; prefilled from FPL"
          },
          "free_transfers": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 0.0,
            "title": "Free Transfers",
            "default": 1
          },
          "max_transfers": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 0.0,
            "title": "Max Transfers",
            "default": 5
          },
          "horizon": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Horizon",
            "default": 5
          },
          "target": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array",
                "maxItems": 15,
                "minItems": 15
              },
              {
                "type": "null"
              }
            ],
            "title": "Target",
            "description": "target squad (default: best team)"
          },
          "path": {
            "type": "boolean",
            "title": "Path",
            "description": "also plan the path to the target squad",
            "default": true
          }
        },
        "type": "object",
        "title": "TransferIn"
      },
      "TransferOptionOut": {
        "properties": {
          "n_transfers": {
            "type": "integer",
            "title": "N Transfers"
          },
          "moves": {
            "items": {
              "$ref": "#/components/schemas/Move"
            },
            "type": "array",
            "title": "Moves"
          },
          "hits": {
            "type": "integer",
            "title": "Hits"
          },
          "expected_points": {
            "type": "number",
            "title": "Expected Points"
          },
          "net_gain": {
            "type": "number",
            "title": "Net Gain"
          },
          "bank_after": {
            "type": "integer",
            "title": "Bank After"
          },
          "recommended": {
            "type": "boolean",
            "title": "Recommended"
          }
        },
        "type": "object",
        "required": [
          "n_transfers",
          "moves",
          "hits",
          "expected_points",
          "net_gain",
          "bank_after",
          "recommended"
        ],
        "title": "TransferOptionOut"
      },
      "TransferOut": {
        "properties": {
          "gw": {
            "type": "integer",
            "title": "Gw"
          },
          "current": {
            "$ref": "#/components/schemas/Squad"
          },
          "bank": {
            "type": "integer",
            "title": "Bank"
          },
          "free_transfers": {
            "type": "integer",
            "title": "Free Transfers"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/TransferOptionOut"
            },
            "type": "array",
            "title": "Options"
          },
          "recommended": {
            "type": "integer",
            "title": "Recommended"
          },
          "path": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PathOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "caveat": {
            "type": "string",
            "title": "Caveat"
          }
        },
        "type": "object",
        "required": [
          "gw",
          "current",
          "bank",
          "free_transfers",
          "options",
          "recommended",
          "path",
          "caveat"
        ],
        "title": "TransferOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}